requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
discord.py==2.4.0
fonttools==4.55.0
frozenlist==1.5.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.3
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pillow==11.0.0
propcache==0.2.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.3
yarl==1.18.0
